requirements.txt
amqp==5.1.1
asgiref==3.7.2
billiard==4.1.0
celery
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cmake==3.27.5
decorator==4.4.2
Django==3.2.13
django-background-tasks==1.2.5
django-compat==1.0.15
django-cors-headers
django-filter==23.3
djangorestframework==3.14.0
drf-yasg==1.21.7
filelock
greenlet
gunicorn==21.2.0
idna==3.4
imageio
imageio-ffmpeg
inflection
Jinja2==3.1.2
kombu
lit
llvmlite
Markdown==3.4.4
MarkupSafe==2.1.3
more-itertools
moviepy
mpmath
networkx
numba
numpy
nvidia-cublas-cu11
nvidia-cuda-cupti-cu11
nvidia-cuda-nvrtc-cu11
nvidia-cuda-runtime-cu11
nvidia-cudnn-cu11
nvidia-cufft-cu11
nvidia-curand-cu11
nvidia-cusolver-cu11
nvidia-cusparse-cu11
nvidia-nccl-cu11
nvidia-nvtx-cu11
openai-whisper
packaging
Pillow
proglog
prompt-toolkit
pydub
python-dateutil
pytz
PyYAML
regex
requests
rest-api-payload
six
SQLAlchemy
sqlparse
sympy
# tiktoken
torch
tqdm
triton
typing_extensions
tzdata
uritemplate
urllib3
uuid
vine
wcwidth
whitenoise
